refactor: clean up structure of files
